Update style conflict from merge commit
